Replace servo-freetype-sys depedency
This replaces the servo-freetype-sys depedency with freetype-sys from
the piston project.

The goal is to standardize on a singular system library, to make sure no
conflicts come up when trying to combine piston's and servo's freetype
and fontconfig libraries.

While it would be possible to instead deprecate freetype-sys in favor of
servo-freetype-sys, I think it makes sense to standardize on the library
that has the most generic name and is easy to search for.